Fix af import fejl
